newest changes to Arduino added

+++ Arduino/TCOSerialTestRead/TCOSerialTestRead.ino
@@ -9,7 +9,7 @@ void setup(){
 void loop(){
   if(tcoSerial.readSerial()){
     // set global state
-    int i = 0;
+    byte i = 0;
     while (i < tcoSerial.readIndex){ // TODO: Write logic for signals
       Serial.write(tcoSerial.readQueue[i++]);
     }
